Fix connection timeout issue
The first connection to PP API service (any service) will always fail
due to deadlock in AndroidServiceBinder.

The problem is that we hold the lock while waiting for connection to be
established. Note: we hold the lock for the whole helper function.
http://shortn/_pSOGj6lmQv

On the other hand, we also try to acquire the lock on the
Binder.onServiceConnected which is on the binder thread.
http://shortn/_Hhv7MWzbnx

These 2 threads cause a deadlock until the CountDownLatch timeout.

Test: atest AdServicesServiceCoreUnitTests and manual test E2E
Bug: 225862831
Change-Id: I9bb5d5c67492951e1766e016f39b6fa907a2d9b2