Added a Response Utility class, which contains common status codes and functions that will be used by CA, ReportImpression, and AdSelection APIs. Modified the CA,ReportImpression, and AdSelection Response classes to use this utility class instead of the existing implementation.
In addition, created a FledgeErrorResponse class, meant to serve as a
generic error response object for Fledge  API's in order to cut down on
duplicated code.

In addition, modified the AdSelection, CustomAudience, and
ReportImpression callback interfaces to contain
two functions instead of one, an onSuccess and an onFailure. Modified
the reportImpression,runAdAuction,joinCustomAudience, and
leaveCustomAudience function in both managers to use the new
generic response class as well as the new callback functions.

Test: Updated Unit tests to reflect the usage of the utility class, and
added a new unit test for FledgeErrorResponse

Bug: 224865403

Change-Id: Ib655a43b0a276bb3cdac207dd5f4cd47280482ec