Reconcile missing sdk data dirs and subdirs on user unlock
If an app is installed when device is locked or if app starts consuming
new sdk after an update but we miss the broadcast for any reason, sdk
data can get out of sync.

When user unlocks, we now iterate over all installed packages, select
the ones that uses sdk and check if sdk data dirs for them is as
expected or not. If unexpected, then we reconcile.

Bug: 217541689
Bug: 226325811
Test: atst SdkSandboxStorageHostTest
Change-Id: Ie96da419ad932399d05ba7ec8e9b0dc5523d763b