Added broadcast receiver in system server.
The broadcast receiver is registered in a context in the SystemService
AdServicesManagerService.  The broadcast receiver is registered after
the device boots up to avoid receiving broadcasts for changes to system packages
and attempting to call the API before the AdServices services have been registered.

The broadcast receiver listens for only uninstall broadcasts  and calls into
a new AIDL API in the AdServices process to tell the AdServices process about the
app uninstallation.

On uninstall the AdServices process will only delete app records from
measurement tables.

Test: manually uninstalled app to check broadcast.
Change-Id: Id3731fb2941bbe4aac0f4e8faf54bfbf2998ed20

Change-Id: Ib8144c3cab889e9e26ae54de05c29f2fda9b6bfe