Allow apps to send data to SDKs in the sandbox
Expose sendData API to apps to send data to the SDK. A callback can be
used by the SDK to indicate whether the the processing of data receieved
was successful or not. To enable bidirectional communication from the
SDK back to the app, the SDK can pass a bundle to the callback as well.

Bug: 235239061
Test: atest SdkSandboxManagerTest
Change-Id: Ie21ff3fe78d73d6415bbf14d542439f70afabc04