SdkSandboxStorageHostTest: remove useless code for emulated FBE
SdkSandboxStorageHostTest contains code that tries to enable emulated
FBE.  However, it only runs if emulated FBE is already enabled, which
makes it useless.  Moreover, both of the test methods that use this code
are skipped on devices that aren't using native FBE, which makes it
doubly useless.

And in any case, emulated FBE is obsolete and is planned to be removed.

For all these reasons, remove this unneeded code.

Test: atest SdkSandboxStorageHostTest # on Cuttlefish
      Note, testSdkDataPackageDirectory_IsClearedOnFreeCache fails both
      before and after this change.
Change-Id: If60d841bde37e83822b86225961d3070fcb56dbf