Handle failure when creating isolate
Add a call to close the connection to the WebView process when PPAPI is
terminated and when any call to create a new isolate fails because the
WebView instance was terminated.

The implemented solution lets the current call fails and closes the
currently active singleton AwJsIsolate instance enabling future call to
succeed. It is responsibility of the caller to retry the call.

Test: JSScriptEngineTest
Bug: 234368175
Change-Id: I67d4dafce1f521676356c1e39015ea4180f31539