Ignore testSanboxCannotAccess_MediaStoreApi.
This test worked because app visibility restrictions prevented us from
resolving the MediaProvider. Now that we have proper app visibility
restrictions in place, we are actually able to resolve the MediaProvider
content provider, and this query would have worked (after fixing the
mime type). For now, ignore the test until we decide what we want to do.

Bug: 234535387
Bug: 234563287
Test: atest SdkSandboxRestrictionsTest
Change-Id: Id93d1f88db26afbf53a22291151b96ae6fdd621f