Create a separate callback for requestSurfacePackage
Avoid using the same callback for both loadSdk and
requestSurfacePackage APIs by creating a new one -
RequestSurfacePackageCallback that the app can use instead to send to
the sandbox manager service. For communication between the sandbox
manager service and the sandbox, another callback
RequestSurfacePackageFromSdkCallback using which the sdk sends back the
view.

Bug: 233620467
Test: atest SdkSandboxManagerServiceUnitTest
Test: atest SdkSandboxManagerTest
Change-Id: I88fa9d43ca9b765b0f9ae01e79345fffc6f754bf