[AdServices] SandboxSdkProvider is a ContextWrapper
SandboxSdkProvider is now context wrapper. This is per API council
feedback.

Also, hide SandboxedSdkContext.
None of the newly added methods in this class were exposed, and the
class was showing up as an empty class in the interface. This removes
the @hide annotation from individual methods and adds it to the class.

Bug: b/238724532, b/238009263
Test: atest
Change-Id: I37a175817f86291586f3ec8bc8a21fbff1b38b92