Inform apps of sandbox death through callbacks
Apps can use SdkSandboxLifecycleCallback to receive notifications about
the death of the sandbox after registering the callback via
registerSdkSandboxLifecycleCallback(). If the sandbox has not been
created yet (for e.g. if the app tries to register the callback before
the sandbox is created), the callback is queued until the sandbox is
created.

Bug: 238761711
Bug: 240437810
Test: atest SdkSandboxManagerServiceUnitTest
Test: atest SdkSandboxManagerTest
Change-Id: I5c6d177f643d167a84dca1ab454e967dbcb50992