Add unloadSdk API
This new API allows client apps to unload SDKs that were
previously loaded by them. The SDK provider will receive a
beforeUnloadSdk call before this occurs. When unloadSdk is called,
all references to the loaded SDK will be deleted. If there
are no more SDKs loaded by the app, then the sandbox is stopped.

Also cleaned up some logic for handling app death. In the case
that an app loads multiple SDKs, only one DeathReceiver will be
maintained for that caller. This ensures that the cleanup logic
will only occur once when the app dies.

Test: atest SdkSandboxManagerTests
Test: Manual test app. Load and unload the SDK, and note
      that the sandbox is killed when the SDK is unloaded. Note
      that the SDK provider receives an onUnloadSdk call.
Bug: 238726140
Change-Id: I1a430a922649fc708de6c484340db435c4c5bd45