Inform apps when calling a non-existent sandbox
Create a specific error code to send apps when a call to the sandbox
fails due the sandbox process not being alive. DeadObjectException is
caught to detect if the sandbox is not alive when making a call.

If the sandbox dies after an async call is made, it is not detected and
callbacks from the sandbox back to the app is not invoked. To fix this,
a list of callbacks which have not been invoked yet are maintained, and
the error code is sent back through the calllbacks if the sandbox dies.

Bug: 238761711
Bug: 238602704
Bug: 241527772
Test: atest SdkSandboxManagerTest
Test: atest SdkSandboxManagerServiceUnitTest

Change-Id: Ica6ffc0a2c3e4f6f2ce201b0fa9eb93f9631f97c