Add support for sycing string keys from app to sandbox
In order to sync data from app to sandbox, we need APIs to communicate
from app to SdkSandboxManagerService and then from
SdkSandboxManagerService to SdkSandboxService.

This CL sets up the foundation for such communication. It's not perfect
yet and many functionalities are still missing, but it's a start. We
will iterate over this.

Added unit tests for various new APIs created, but none of the logic has
been integrated with each other yet. Integration logic and e2e tests
will be added separately.

Bug: 239403323
Test: atest SdkSandboxUnitTests
Test: atest SdkSdkSandboxFrameworkUnitTests
Test: atest SdkSandboxManagerServiceUnitTests
Change-Id: I30b1dbe59da383c14ee5e56d58d43b691fcd9207