Add support for syncing removal of keys from SharedPreferences
Keys can be removed by either calling remove() or setting the value of
the key to null on SharedPreference.

Removal of key is conveyed through SharedPreferencesUpdate by including
it as one of the key in getKeysInUpdate() list and excluding it from
getData() bundle.

Additionally, key can be removed by clearing all the keys via clear()
api on editor. Added support for that too.

Bug: 239403323
Test: atest SdkSdkSandboxFrameworkUnitTests
Change-Id: Ic3656bcb2e4d6113fcaf76c29ba0e4da407185e6