Add support for callbacks, error reporting and stopping the sync 1/2
While syncing, it is possible to face errors such sandbox not being
available or failure to write in persistent disk on sandbox side. We
need callbacks to report back errors to the user.

There are two callbacks introduced in this CL:
  - one for the user which can notify user about start, stop and error.
  - one for internal communication, which lets SyncManager know about
    successful writing inside sandbox, errors and sandbox start event.

Using the callbacks, we can now allow users to start syncing their keys
before sandbox start. SyncManager is now capable of waiting for
notification for SandboxStart and starting the sync.

For all kinds of other errors, we simply notify the user and break the
sync so that they can retry or act knowingly that sync isn't working.

User may also want to stop syncing the keys. That should also be
possible now.

Only one set of keys can be synced at any given moment. If user wants to
sync a different set of keys they will need to stop the sync manually
first and start with new set of keys.

Old keys will not be removed, so there is a potential of storage leak if
user is not careful.

This CL only contains changes for SdkSandboxManager. Related changes for
SdkSandbox and SdkSandboxManagerService is split out into a separate CL
to keep this one small!

Bug: 239403323
Test: atest SdkSdkSandboxFrameworkUnitTests
Change-Id: Ib771ffdafe18d8e8fb0a4fb34bb64be19c3555a8