Add support for callbacks, error reporting and stopping the sync 2/2
The previous CL mostly focused on logics in SdkSandboxManager. This CL
takes care of SdkSandboxManagerService and SdkSandbox.

When SdkSandboxManagerService receievs call for sync, it needs to
forward the call to Sandbox. On the event sandbox is not available yet,
manager needs to cache the callback and call it later when Sandbox
starts. Otherwise, SyncManager has no way to know when sandbox has
started, other than polling frequently.

On the sandbox side, on writing the data to SharedPreferences
successfully, we call onSuccess.

Bug: 239403323
Test: atest SdkSandboxManagerServiceUnitTests SdkSandboxUnitTests
Change-Id: Ie825fa85cc9812bd0e4e2aa10d9d6c9dc61e7f5e