Integrate SyncManager with SdkSandboxManager
We are exposing three new APIs in this CL:

1. addSyncedSharedPreferencesKeys
2. removeSyncedSharedPreferencesKeys
3. getSyncedSharedPreferencesSyncKeys

The idea is to provide a feature using which app or some sdk in
app can keep some values from default SharedPreferences synced with
Sandbox.

One of the use cases we forsee for this feature is for keeping consent
related data synced between app and sandbox. Certain CMP SDKs collect
and store data in default SharedPreferences which the Ad SDKs then
consume. With SdkSandbox isolating SDKs from the app, they now need a
way to read data from app's default SharedPreferences, which is where
this feature comes in.

Details regarding how the api works have been added in their Javadoc.

This CL also integrates the feature with existing flow. Integration was
as simple as just forwarding the calls for start/stop to SyncManager.

Added logging for tracking sync lifecycle. Update name of the tag to
SdkSandboxSyncManager to isolate sync related logging from other logs.

Added e2e tests to ensure integration is working as intented.

Bug: 239403323
Test: atest SdkSandboxStorageHostTest
Change-Id: Ib0e396c10f7820e1cc50253ecb7645e560737351