Add support for syncing non-string values
SharedPreference supports 6 kinds of values: Boolean, Float, Int, Long,
String and StringSet. There was no api to put a StringSet in Bundle, so
passed it as an ArrayList.

There is also no api to figure out what value has been stored against
a key, so we will ask the user to provide us with the type when they
share the list of keys. If specified type is not found, we ignore the
key for now. In future, when we implement error reporting, we will send
error back to the user.

There is still no way to pass information about type of key via bundle.
On next CL, we will create separate parcelable class specific to
encapsulate SharedPreferences update information. With that, we will
be able to implement logic for syncing data to Sandbox.

Bug: 239403323
Test: atest SdkSdkSandboxFrameworkUnitTests
Change-Id: I7c8f43c110696481c7f0876fe656b5d33ad59d90