Fix testDuplicateLoadingFails()
The test tries to load the same SDK before getting the result of loading
it the first time, which can cause the test to fail at times. A fix is
added where it is first checked if loading the SDK was a success or not.

Bug: 242720236
Test: atest
com.android.sdksandbox.SdkSandboxTest#testDuplicateLoadingFails

Change-Id: Ic2456d781dc679fd45d8cbe605ccca1b0a888a94