Set the correct context for ppapi managers when running inside sandbox
Unfortunately the current implementation of SandboxedSdkContext
(extending ContextWrapper) means that by default any Manager class
obtained from SandboxedSdkContext.getSystemService(FooManager.class)
will be created using the base context instead of the
SandboxedSdkContext context.

Fixing this requires a platform change, so for the time being we
introduce the SdkSandboxSystemServiceRegistry workaround which allows us
to register names of system services that should be mutated before
returning to the caller. This allows us to propagate the right
(SandboxedSdkContext) context to the PP APIs related Manager classes.
This change uses TopicsManager as an example.

On top of that this change also fixes the system services caching
problem inside sdk sandbox process. Since the base context of
SandboxedSdkContext was application context, the results of the
SandboxedSdkContext.getSystemService calls were effectively cached on
the process level. With this change we are introducing a new
intermediate context (obtained via
Context#createDeviceProtectedStorageContext() that SandboxedSdkContext is
based off. Since Context#createDeviceProtectedStorageContext() will create
a new instance of ContextImpl, this ensures that returned system will be
effectively cached on per-SandboxedSdkContext basis.

Test: presubmit
Bug: 237670590
Change-Id: I32fd5657828cf00b3752adbb841e58f098eb37b1