Handle sandbox death while unloading SDK
If the sandbox dies and unloadSdk() is called after, an exception is
thrown since the SDK is no longer loaded. This can cause the app to
crash if the exception is not caught on the app side. To handle this, in
case the sandbox is not available at all when unloadSdk() is called, the
operation is no-op.

Bug: 243028191
Test: atest SdkSandboxManagerServiceUnitTest
Change-Id: I38f456db0bea092da9b67771d6f56a405b469bc0