Fix flaky SdkSandboxTest
Test flakes in SdkSandboxTest might be occurring because
latch.countDown() in onSurfacePackageReady() is called before setting
the value of the surfacePackage variable. The thread running the test
might then read the non-updated value of the variable first. This issue
is also present in other callbacks.

This CL introduces a potential fix.

Bug: 242830758
Bug: 242830817
Bug: 242830208
Bug: 242830229
Bug: 242830113
Test: atest SdkSandboxUnitTests
Change-Id: I910f0833e97cf297dc923243550f53e8f4d85cd3