Convert SharedPreferencesSyncManager into a singleton
This class is responsible for syncing a set of keys to sandbox. If there
are multiple instances of this class, then it is possible to end up
syncing mutliple set of keys, with the sync mechanism getting on each
others toes. This is not something we want.

Making this class ensures there is only one sync mechanism running with
complete view of all the keys being synced.

Bug: 239403323
Test: atest SdkSdkSandboxFrameworkUnitTests
Change-Id: If4dea8aca09b8826d9cede04700cfaca9d947f6a