Convert startSharedPreferencesSync to incremental API
Setting keys as a list can lead to race conditions where multiple
entities tries to set keys at the same time.

An incremental approach (addKeys/removeKeys) would suit better than
start/stop API here.

User can now call this API multiple times and we have removed a lot of
the restrictions.

In order to maintain a pool of keys that can add new keys in it, I
converted mKeysToSync to a map of key->startSharedPreferencesKey. That
made mTypeOfKey redudandent, so I removed it.

Bug: 239403323
Test: atest SdkSdkSandboxFrameworkUnitTests
Change-Id: Ib3cf9376b08d1be9432be2ccca6301d7c7e29708