Remove callback from addSharedPreferencesSync API
The callback was used to notify the user that:

1) Sync has started.
2) Sync has stopped.
3) There was an error.

The sync should be as transparent as possible such that user doesn't
have to worry about it all.

With incremental API design, the sync should always run. Even when
sandbox is unavailable, the user doesn't need to know that sync is
paused. From users point of view, they have provided the key and
platform should sync seamlessly.

So #1 and #2 is unnecessary for them.

If an error occurs during sync, #3 would notify the user. But the user
doesn't have action to take. They can remove the key from sync pool, but
due to error, sync has already stopped, making that action redundant.

Bug: 239403323
Test: atest SdkSdkSandboxFrameworkUnitTests
Change-Id: I233c5f2f8c3003b8d88cafccea68cc6525e3f429