Reduce the scope of ISharedPreferencesSyncCallback
We no longer need onSuccess, as client is not interested in knowing when
we have started sync or written data on Sandbox. So removed logic for
forwarding the callback to sandbox.

Previously, we were registering the update listener on bulk sync success
only. This was unnecessary. This class is thread-safe so there is no
risk in running ChangeListener and Bulk-Sync at the same time.

Also, there is also no need for sending unique callbacks to
SdkSandboxManagerService. We are only interested in knowing if sandbox
is available or not. Converted the callback to a singleton.

Bug: 239403323
Test: atest SdkSdkSandboxFrameworkUnitTests
Change-Id: Ifa71aae4e529358638d040ecef3cb88419e33cb3