Disable sandbox if Webview provider cannot be resolved
Adds a check to SdkSandboxManagerService to determine whether the
sandbox should be disabled on the device. This is done by trying
to resolve the Webview provider when the first SDK is loaded.
If the Webview provider cannot be found, the sandbox is disabled.

This change does not perform any checking of the sandbox state when
calling SdkSandboxManager APIs. This logic will be added later.

Test: atest
Test: flash device with and without Webview fix, check sandbox state.
Bug: 238874537
Change-Id: I26e75e36482731582d1d740c7673d3739df52ae3