Lazily create Classifiers
Currently all Topics related APIs and workflow will create the
TopicsWorker which will in turn create CacheManager, EpochManager,
ClassifierManager. The ClassifierManager will create OnDeviceClassifier
and PrecomputedClassifier.
However in normal flow, we don't need Classifiers. Creating them is
heavy and cost a lot of memory. In this cl, we lazily create them when
it's first used and therefore save memory and latency.

Bug: 243191239
Test: presubmit
Change-Id: I80c029d5525974aa56938692cce7e9f84db67e0d