Add SDK sandbox killswitch
Adds a new property to the sdk_sandbox namespace which can be used
to remotely disable the sandbox. When this switch becomes enabled,
all existing sandboxes will be killed. This killswitch will block
SDKs from being loaded in a future change. The killswitch is disabled
by default, and cannot change from disabled -> enabled -> disabled
in a single boot.

Also added a shell command to force enable the sandbox, or to reset
the sandbox state.

Test: atest SdkSandboxManagerServiceUnitTests
Test: atest SdkSandboxLifecycleHostTest
Test: Manual: load SDK and request surface package, trigger killswitch
      from command line, note that sandbox is killed and surface package
      click events no longer work.
Bug: 238874537
Change-Id: Ibe97092cc9fe2d6503b20dbb37d0d1bd58dd95f0