Fix attribution job sometimes scheduling
After attribution job finished, rescheduling was not consistent. Sometimes rescheduling was successful, sometimes it wasn't. The method call to jobFinish is asyncronous. Rescheduling only takes place if the job finished, therefore if jobFinish is called and the system takes too long to receive the information, when trying to reschedule the system might think the job is still in a pending state instead of finished. Therefore, forcing scheduling.

Test: atest AttributionJobServiceTest
Change-Id: I6d973a2bdeca3bf0a5d82d7f1315d52d63d10e21