Fail loadSdk when sandbox disabled
Adds a new error code that will be returned in a
LoadSdkException. This error code will be returned when
the sandbox is disabled, either due to the killswitch
being enabled or the Webview visibility patch not existing
on the device.

Also adds a target preparer to existing tests to ensure that
tests can still run on devices without the Webview patch. This
is done by force-enabling the sandbox for the tests.

The 'start' shell command will also fail if the sandbox is
disabled. Added unit tests for this.

Bug: 243399324
Test: atest SdkSandboxManagerDisabledTest
Test: atest SdkSandboxManagerServiceUnitTests
Change-Id: I82444903ce833bf9c6821d64408a13f1d485ac28