Fix bug in test verifing sdk root directory is deleted on user removal
Test was failing on bug fix since user data deletion is asynchronous.
Added a waiting logic. It roughly takes an extra second for user data to
be deleted, but added a loop for 30 seconds with 1 second breakout
checks since that would be less flaky.

Renamed the test since it's related to sdk root directory, not package
level directory.

Bug: 245943997
Test: atest testSdkDataRootDirectory_IsDestroyedOnUserDeletion
Change-Id: I1356ef65ad9cce2b6a5fadf57214d621a207d8d9