Unbind all sandboxes with the same UID
When a sandbox is stopped, it is first unbound and then killed by UID.
If there are several apps with a shared uid and we are stopping a
sandbox of one app, all sandboxes with the same uid get killed while
only the one sandbox gets unbound. This would mean that the other
sandboxes are prone to being respawned again. This could cause erroneous
behaviour if the other apps die before the sandbox gets respawned.
Additionally, there is no use of respawning a sandbox since all the SDKs
loaded into it are already lost.

Add a change so that all sandboxes with the same uid are unbound before
calling killUid().

Bug: 246254104
Bug: 230839879
Test: atest SdkSandboxLifecycleHostTest
Change-Id: I2b48b39894320ac2fd4d72b2b3c0ed0f79f818ed