Add video option to sandbox manual test app
Adds the option to create a VideoView inside the SDK
sandbox. This is done by passing a String extra to the
main activity from the command line, signifying the video
URL to be loaded. This will allow a "Play Video" button to
be clicked after the SDK has loaded, which will create a VideoView
for the passed URL.

Test: adb shell am start -n com.android.sdksandboxclient/.MainActivity --es "video-url" "<url>"
Bug: 246928295
Change-Id: I4e9f783d54a6ee9f730a33f43a82e044af074dbc