Introduce a new optional parameter mRecordObservation
We add a new parameter mRecordObservation in GetTopicsRequest and getTopicsParam to track if caller want to  record that caller has observed the Topics of the host app or not. This will be used to determine if the caller can receive the topic in next epoch. The default value is "true", if caller doesn't set mRecordObservation parameter. Beta 1 clients using the AdvertisingTopicsClient will not need to make any code change.
Chrome's equivalence API: https://screenshot.googleplex.com/4dkKwvwvEyapM3A

Test: atest AdServicesFrameworkUnitTests

Bug: 246979195
Change-Id: Ib96823e3ebc658e147bd937c4b602ec0ecf873fe