Fixes READ_DEVICE_CONFIG for measurement jobs test
Bug: 247696929
The issue was that PhFlags were getting called on AdServicesConfig,
because it was not getting mocked. The error is not consistent as the
call happens on a separate thread. The error was not spotted right away
as the rule TestableDeviceConfig makes reading configs from device work,
but it looks it is not that reliable. The change includes removing
TestableDeviceConfig rule so if any new flag is introduced in the code
flow, it might be spotted right away and the necessary changes can be
applied.

Test: atest com.android.adservices.service.measurement
Change-Id: I7b1ce63f6d03d724d8bca4d030fe5b42f0daa047