Use interface returned by loadSdk to interact with sdk in Storage Tests
Relying on requestSurface package is brittle and cumbersome. Using
interface allows us to pass parameters and get results back seamlessly.

Exception handling is still a bit cumbersome. We can't propagate checked
exceptions via binder. Only a subset of RuntimeException is supported.
We are using IllegalStateException as a substitue to carry the real
exception back to the test app.

Bug: 242678799
Test: atest SdkSandboxStorageHostTest
Change-Id: Ie20d7aa0b7c18c2bc503ce377a7260e3a3eae257