Keep track of pending SDK loads
Track loadSdk() calls that have not completed yet in order to prevent
the app from loading the SDK while the previous load is in progress.

Previously, mAppAndRemoteSdkLinks was used to keep track of this and
explicit clean up of the remote sdk link was required in some cases.
Instead, introduce a new set only for pending calls, which gets
automatically cleaned up on the completion of a loadSdk() call or on
sandbox death.

Bug: 248292105
Test: atest
SdkSandboxManagerServiceUnitTest#testLoadSdk_firstLoadPending_errorOnLoadAgainRequest

Change-Id: Ia1063adbeedc47fdec292466bbab73a2b8bbf4ca