Create API in SdkSandboxController for fetching synced data
Data synced from client app using addSyncedSharedPreferencesKeys should
be accessible by the sdks inside sandbox.

A separate API is exposed since the synced data is kept in a separate
location. By keeping data in separate location, we prevent client data
from mixing with sdk data. This allows us to delete the synced data on
sandbox restart, preventing data leak when client removes key for
syncing when sandbox is not alive.

Due to b/248214708, we were not able to keep the synced data in a
separate internal storage area. We have instead opted to store the data
in named a SharedPreferences within the shared storage. The name starts
with "com.android.sdksandbox", highlighting it's for use by sdksandbox
and should prevent accidental misuse.

Bug: 237410689
Test: atest SdkSandboxStorageHostTest
Change-Id: I9384c23a24604369c92fa9717a4c3d615d79cd33