Improve user creation/deletion logic in storage test
User deletion is an async process. Time taken to remove user can vary.
We currently wait for 30 seconds every time user is removed, but on
occasions that may not be enough. This makes the test flaky.

When removing users, we don't always care if the user data has been
deleted or not. For cases when we don't care, we should not wait for
user data to delete. Since system creates new user with increasing user
id in the same boot, there should not be any conflict.

Also, increased the timeout to 120 seconds for user deletion which is
consistent with the time we wait for user creation.

Bug: 249306347
Test: atest SdkSandboxStorageHostTest
Change-Id: I58317b2bc4878a1227a73e29c70ae389e986639b