Introduce a new optional parameter mRecordObservation.
We add a new parameter mRecordObservation in GetTopicsRequest and getTopicsParam to track if caller want to record that caller has observed the Topics of the host app or not. This will be used to determine if the caller can receive the topic in next epoch. The default value is "true", if caller doesn't set mRecordObservation parameter. Beta 1 clients using the AdvertisingTopicsClient will not need to make any code change.
Chrome's equivalence API: https://screenshot.googleplex.com/4dkKwvwvEyapM3A

Bug: 246979195

Test: atest  AdServicesServiceCoreUnitTests AdServicesFrameworkUnitTests AdServicesEndToEndTests CtsAdServicesEndToEndTests CtsAdServicesDeviceTestCases AdServicesApkUnitTests CtsSandboxedTopicsManagerTests --no-bazel-mode
Change-Id: I8415c55b4ea57710cebb3ff96786ba8c449b9ade