Allow test app to be moved to a different volume
The test app has no code and no oat directory is created for it. This
fails a PackageManager check that allows apps to be moved only if the
oat directory is present. As a result, tests that move the app to
another volume fails.

Add some placeholder code in the test app to prevent this.

Bug: 250200859
Test: atest SdkSandboxDataIsolationHostTest
Change-Id: I865afcd8ca178f215f1b6ce5926affcc19248153