Fixes rateLimiter no initial burst
Bug: 253007573.
The RateLimiter burst mechanism was not getting triggered on the first subsequent calls. This meant that if Throttler was configured with a value of 5 and the first 5 requests arrived simultaneously, the 1st call permit would be granted while the other 4 would be denied. As they are the first 5 requests, they would have to be spread out evenly over 1 second as bursting didn't kick in yet. With these changes, all 5 requests would be granted as per the permits per second configured. Additionally, allow different APIs configure their own permits per second limit

Test: atest AdServicesServiceCoreUnitTests
Change-Id: Idbfb56789e6795354c995f278f40fd8681e5fb3a