Don't throw exceptions if SDK is not loaded
Stop throwing IllegalArgumentException while checking if an SDK is
loaded before interacting with it as the app could crash otherwise. When
calling APIs to interact with the sandbox, the SDK might not be loaded
for various reasons:
* The sandbox could have died after loading.
* The app might have accidentally called requestSurfacePackage() before
  loading of the SDK is complete.
Instead, inform the app that the SDK is simply not loaded.

Additionally, stop checking for whether the sandbox process is available
or not at the start of the interactions as this could cause confusing
behaviour for app developers otherwise:
* If the sandbox had not been brought up at all, the error message would
  have been SANDBOX_NOT_AVAILABLE.
* If the sandbox had been brought up, but some SDK not loaded, the error
  message would have been SDK_NOT_LOADED.
Change such that the behaviour is always consistent - SDK_NOT_LOADED is
always shown.

Bug: 254567923
Bug: 249482251
Test: atest SdkSandboxManagerServiceUnitTest
Test: atest SdkSandboxManagerTest
Change-Id: I7cd91d4ee91af2eadeccaba8287d5a235e23221f