Use SDK binder interface to communicate with SDKs in tests
Data isolation tests use the bundle passed by requestSurfacePackage() to
communicate with SDKs. Use the binder interface returned by the SDK on
loading instead, as it is cleaner.

Additionally, using assert statements lead to the sandbox dying and
causing a DeadObjectException app-side. Throw IllegalStateException
instead to properly carry the exception.

Bug: 251771403
Test: atest SdkSandboxDataIsolationHostTest
Change-Id: I3d1a031ab7355c1f9a94cca3a7e17d7d7820658b