Add buyer fields to CustomAudienceStats
The current CustomAudienceStats object only counts on owners and should
support counting on buyers at the same time.  This will be used later
when per-buyer limits are introduced and when reporting the number of
distinct buyers in the custom audience database.

Also moved the CustomAudienceStats to an AutoValue builder pattern.

Test: cd packages/modules/AdServices/adservices && atest *AdServices*
Change-Id: I40d3bc510d8df475d5b0cbef57403ed87f334d98