Add custom test instances to SDK Sandbox test utils
The WebView team is going to be sharing tests between the
SDK Sandbox and Activity based environments.

In order to support this, they will be defining a shared
test suite that will need to be extended in each
environment. The problem is that the in SDK tests
already need to extend SdkSandboxTestScenarioRunner.

To get around this, allowing SDK test runners to specify
custom test instances that will contain tests.

Test: Tested using shared SDK tests I have locally.
Bug: 255745882
Change-Id: Ifc42cb711e267e670bfa5b0264e79fbaf0004c0c