Prefactoring Tests to allow to Mock Flags
We're working on ag/20270887 to enable a flag to guard DB schema change. This requires to introduce FlagFactory in DBHelper. It results in many measurement tests to fail due to Device_config access issue to read production flag in test.

In order to split the fix from actual logic, I raise this CL to prefactoring affected tests. I fix some of them by changing db to test db, and fix others by mocking the flag directly.

Bug: 256028076
Test: atest AdServicesServiceCoreUnitTests AdServicesApkUITests
Change-Id: I6d9a62c7e5caba9f705820bdb134aad33a683c48