Add test author binders to sdk sandbox test utils
There are situations where test authors will
need to invoke behavior outside of the SDK
runtime but have it initiated from inside
the SDK runtime tests.

For example, a test author may want to
communicate with the instrumented context
of the app in control of it (for example
to perform a touch event).

By allowing the even to be invoked from
within the test SDK, this means tests
could be shared with non SDK runtime tests
and it also means it would be easier
to read test code because it could all
be within the same location.

Test: Added a custom binder to a test
locally and confirmed I could invoke methods from
it. Also confirmed that the tests would execute
properly without a binder provided.
Bug: 255745882

Change-Id: I05b7e0e2b409fba4d5e821b7ebd719fb60dd961a