Add docs and example for sdksandbox testscenario
Now that we've added sdk test utils, we are adding
documentation so that anyone can make use of these and
work on them in the future.

The README.md contains general tips on writing JUnit
test suites with these tools. Also adding an
example project that could both be useful for development
purposes and understanding how to create a new test suite.

Test: atest SdkSandboxTestUtilsExampleTest
Bug: 257962733
Change-Id: I57a6833be88bdc63d125c9819ca36828e64f2c4b