Remove UiThreadTest annotation check from Sdk test utils
This was originally added to make it easier to share
tests for webkit cts tests. This won't be necessary
because the webkit tests are rather going to
transition to use a runOnUI method instead of relying
on this annotation.

The AndroidJUnit4 runner does not observe this annotation
either so we should just drop it and use methods to run
on the UI thread.

Test: atest CtsSdkSandboxWebkitTestCases
Bug: 241125957
Change-Id: I7f7625bb9cef816c22df7771b3e64c24f5a80130