Make the parallelisation of CA bidding configurable
We made a change that executes the bids for buyers in parallel but for
each buyer the CAs were bid seuqnetially. This had a hit on latency.
This new implementation breaks the bidding per buyer to in
parallelizable chunks, where each chunk can run in parallel. While each
chunk of work bids on CAs in sequence.

So this hybrid approach gives:
- Perf improvements of parallelization
- benefits of sequential runs for cache hits
- Wasting minimal resources by only terminating at max no parallel
  chunks

Configurability: The number of parallel chunks for each buyer is
configurable

Test: Unit and end to end tests
Bug: 255431848

Change-Id: Iecaa7fac808d71403737b437032cc75ae6190a1d